backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
pydantic==2.9.2
python-dotenv==1.0.1
scikit-learn==1.2.2
joblib==1.4.2
psycopg2-binary==2.9.9
google-genai>=1.0.0
